+++ src/main/java/com/cafe/crm/models/menu/Ingredients.java
@@ -27,7 +27,7 @@ public class Ingredients extends BaseEntity {
 	private String dimension;
 
 	@NotNull(message = "Заполните поле")
-	private Integer amount = 0;
+	private Double amount = 0D;
 
 	private Double price = 0D;
 
@@ -35,14 +35,14 @@ public class Ingredients extends BaseEntity {
 
 	}
 
-	public Ingredients(String name, String dimension, Integer amount, Double price) {
+	public Ingredients(String name, String dimension, Double amount, Double price) {
 		this.name = name;
 		this.dimension = dimension;
 		this.amount = amount;
 		this.price = price;
 	}
 
-	public Ingredients(Long id, String name, String dimension, Integer amount) {
+	public Ingredients(Long id, String name, String dimension, Double amount) {
 		this.id = id;
 		this.name = name;
 		this.dimension = dimension;
@@ -74,11 +74,11 @@ public class Ingredients extends BaseEntity {
 	}
 
 
-	public Integer getAmount() {
+	public Double getAmount() {
 		return amount;
 	}
 
-	public void setAmount(Integer amount) {
+	public void setAmount(Double amount) {
 		this.amount = amount;
 	}
